requirements.txt
# Telegram Bot Framework
aiogram==3.3.0

# Database
aiosqlite==0.19.0
sqlalchemy==2.0.23
alembic==1.13.0

# Google Sheets - using public CSV access (no API credentials needed)
# No additional packages required - uses built-in csv and aiohttp

# Yandex GPT API
yandexcloud

# OpenAI (альтернатива Yandex GPT)
openai

# AmoCRM API
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Date/time utilities
python-dateutil==2.8.2

# HTTP client
aiohttp==3.9.3

# Logging
loguru==0.7.2

# JSON handling
orjson==3.9.10

# PDF generation
reportlab==4.0.7
markdown==3.5.1

# Payments (uuid is built-in Python module)

# CRM API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.1
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0